A3/app/requirements.txt
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.3
typing-extensions>=4.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.0.2
flask-htmx==0.3.1
gunicorn==21.2.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
requests==2.31.0
six==1.16.0
urllib3==2.0.4
Werkzeug==2.3.6
